refactor(personal-assistant): use claude 4 sonnet, organize custom tools, bug fixes, update readme
