+++ 02-samples/05-personal-assistant/.gitignore
@@ -1,0 +1,1 @@
+appointments.db
