+++ 02-samples/05-personal-assistant/calendar_assistant.py
@@ -1,144 +1,12 @@
-import sqlite3
-import uuid
-from datetime import datetime
+import os
 from strands import Agent, tool
 from strands.models import BedrockModel
 from strands_tools import current_time
-import list_appointments
-import update_appointment
+from calendar_tools import create_appointment, get_agenda, list_appointments, update_appointment
+from constants import SESSION_ID
 
-
-@tool
-def create_appointment(date: str, location: str, title: str, description: str) -> str:
-    """
-    Create a new personal appointment in the database.
-
-    Args:
-        date (str): Date and time of the appointment (format: YYYY-MM-DD HH:MM).
-        location (str): Location of the appointment.
-        title (str): Title of the appointment.
-        description (str): Description of the appointment.
-
-    Returns:
-        str: Formatted confirmation of the newly created appointment.
-
-    Raises:
-        ValueError: If the date format is invalid.
-    """
-    # Validate date format
-    try:
-        datetime.strptime(date, "%Y-%m-%d %H:%M")
-    except ValueError:
-        raise ValueError("Date must be in format 'YYYY-MM-DD HH:MM'")
-
-    # Generate a unique ID
-    appointment_id = str(uuid.uuid4())
-
-    conn = sqlite3.connect("appointments.db")
-    cursor = conn.cursor()
-
-    # Create the appointments table if it doesn't exist
-    cursor.execute(
-        """
-    CREATE TABLE IF NOT EXISTS appointments (
-        id TEXT PRIMARY KEY,
-        date TEXT,
-        location TEXT,
-        title TEXT,
-        description TEXT
-    )
-    """
-    )
-
-    cursor.execute(
-        "INSERT INTO appointments (id, date, location, title, description) VALUES (?, ?, ?, ?, ?)",
-        (appointment_id, date, location, title, description),
-    )
-
-    conn.commit()
-    conn.close()
-
-    # Format the confirmation with same style as get_agenda
-    time_part = date.split(" ")[1] if " " in date else "No time specified"
-    date_part = date.split(" ")[0] if " " in date else date
-    confirmation = [
-        "✅ Appointment Created Successfully!",
-        "=====================================",
-        f"📅 Date: {date_part}",
-        f"🕐 Time: {time_part}",
-        f"📍 Location: {location}",
-        f"📝 Title: {title}",
-        f"📄 Description: {description}",
-        f"🆔 ID: {appointment_id}",
-        "",
-        "Your appointment has been saved to your calendar!"
-    ]
-    return "\n".join(confirmation)
-
-
-@tool
-def get_agenda(date: str) -> str:
-    """
-    Retrieve the agenda for a specific day, showing all appointments scheduled for that date.
-
-    Args:
-        date (str): Date to get agenda for (format: YYYY-MM-DD).
-
-    Returns:
-        str: Formatted agenda for the specified date with all appointments.
-
-    Raises:
-        ValueError: If the date format is invalid.
-    """
-    # Validate date format
-    try:
-        datetime.strptime(date, "%Y-%m-%d")
-    except ValueError:
-        raise ValueError("Date must be in format 'YYYY-MM-DD'")
-
-    conn = sqlite3.connect("appointments.db")
-    cursor = conn.cursor()
-
-    # Create the appointments table if it doesn't exist
-    cursor.execute(
-        """
-    CREATE TABLE IF NOT EXISTS appointments (
-        id TEXT PRIMARY KEY,
-        date TEXT,
-        location TEXT,
-        title TEXT,
-        description TEXT
-    )
-    """
-    )
-
-    # Query appointments for the specific date (using LIKE to match the date part)
-    cursor.execute(
-        "SELECT id, date, location, title, description FROM appointments WHERE date LIKE ? ORDER BY date",
-        (f"{date}%",)
-    )
-
-    appointments = cursor.fetchall()
-    conn.close()
-
-    if not appointments:
-        return f"No appointments scheduled for {date}"
-
-    # Format the agenda
-    agenda_lines = [f"📅 Agenda for {date}:", "=" * 30]
-
-    for appointment in appointments:
-        appointment_id, appointment_date, location, title, description = appointment
-        # Extract time from the datetime string
-        time_part = appointment_date.split(" ")[1] if " " in appointment_date else "No time specified"
-
-        agenda_lines.append(f"🕐 {time_part} - {title}")
-        agenda_lines.append(f"   📍 Location: {location}")
-        agenda_lines.append(f"   📝 Description: {description}")
-        agenda_lines.append(f"   🆔 ID: {appointment_id}")
-        agenda_lines.append("")  # Empty line for spacing
-
-    return "\n".join(agenda_lines)
+# Show rich UI for tools in CLI
+os.environ["STRANDS_TOOL_CONSOLE_MODE"] = "enabled"
 
 
 @tool
@@ -151,28 +19,32 @@ def calendar_assistant(query: str) -> str:
     Returns:
         Output from interaction
     """
-    system_prompt = """You are a helpful calendar assistant that specializes in managing my appointments. 
-    You have access to appointment management tools, and can check the current time to help me organize my schedule effectively. 
-    Always provide the appointment id so that I can update it if required"""
-
-    model = BedrockModel(
-        model_id="us.anthropic.claude-3-7-sonnet-20250219-v1:0",
-    )
-
-    agent = Agent(
-        model=model,
-        system_prompt=system_prompt,
-        tools=[
-            current_time,
-            create_appointment,
-            list_appointments,
-            update_appointment,
-            get_agenda
-        ],
-    )
     # Call the agent and return its response
     response = agent(query)
+    print("\n\n")
     return str(response)
+
+
+system_prompt = """You are a helpful calendar assistant that specializes in managing my appointments. 
+You have access to appointment management tools, and can check the current time to help me organize my schedule effectively. 
+Always provide the appointment id so that I can update it if required"""
+
+model = BedrockModel(
+    model_id="us.anthropic.claude-sonnet-4-20250514-v1:0",
+)
+
+agent = Agent(
+    model=model,
+    system_prompt=system_prompt,
+    tools=[
+        current_time,
+        create_appointment,
+        list_appointments,
+        update_appointment,
+        get_agenda
+    ],
+    trace_attributes={"session.id": SESSION_ID},
+)
 
 
 if __name__ == "__main__":
@@ -214,8 +86,6 @@ if __name__ == "__main__":
 
             print("🤖 CalendarBot: ", end="")
             response = calendar_assistant(user_input)
-            print(response)
-            print()
 
         except KeyboardInterrupt:
             print()
