+++ 02-samples/05-personal-assistant/constants.py
@@ -1,0 +1,1 @@
+SESSION_ID = "aws-show-tell-test"
