+++ 02-samples/05-personal-assistant/personal_assistant.py
@@ -1,13 +1,25 @@
+import os
 from strands import Agent
+from strands.models import BedrockModel
 from code_assistant import code_assistant
 from calendar_assistant import calendar_assistant
 from search_assistant import search_assistant
+from constants import SESSION_ID
 
+# Show rich UI for tools in CLI
+os.environ["STRANDS_TOOL_CONSOLE_MODE"] = "enabled"
+
+model = BedrockModel(
+    model_id="us.anthropic.claude-sonnet-4-20250514-v1:0",
+)
 
 personal_assistant_agent = Agent(
-    tools=[code_assistant, calendar_assistant, search_assistant],
+    model=model,
     system_prompt="You are a personal assistant. Use the agents and tools at your disposal to assist the user.",
+    tools=[code_assistant, calendar_assistant, search_assistant],
+    trace_attributes={"session.id": SESSION_ID},
 )
+
 
 if __name__ == "__main__":
     print("=============================================================================")
@@ -55,8 +67,7 @@ if __name__ == "__main__":
 
             print("🤖 PersonalBot: ", end="")
             response = personal_assistant_agent(user_input)
-            print(response)
-            print()
+            print("\n")
 
         except KeyboardInterrupt:
             print("\n")
